backend/requirements.txt
# App
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic-settings==2.3.4
di4injector==0.0.2
slf4py==0.0.4

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# DataBase
SQLAlchemy==2.0.31
mysqlclient==2.2.4 # MySQL
psycopg2-binary==2.9.9  # Postgresql
psycopg2==2.9.9  # Postgresql
redis[hiredis]  # Redis

# Mail
html2text==2024.2.26

# Other
requests==2.31.0  # HTTP Request
python-dotenv==1.0.1  # Environment
pytz==2024.1  # Timezone
PyJWT==2.9.0  # JWT